+++ CHANGES.rst
@@ -5,7 +5,8 @@ PosterKit-O-Mat CHANGES
 
 in progress
 -----------
-
+- lqdn-gafam-campaign: Obtain "lang" and "name" variables from query parameters,
+  e.g. https://examples.posterkit.net/lqdn-gafam-campaign/poster.html?lang=fr&name=google
 
 2018-01-15 0.0.0
 ----------------